pyproject.toml
[tool.poetry]
name = "auto-star"
version = "0.1.0"
description = "Automatically like GitHub repositories with rate limiting"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
apscheduler = "^3.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
auto-star = "auto_star.__main__:main"
